rename file and add users-settings subscription
