+++ bigbluebutton-html5/imports/startup/server/index.js
@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import Langmap from 'langmap';
 import Users from '/imports/api/users';
 import fs from 'fs';
-import './iniSettings';
+import './settings';
 import Logger from './logger';
 import Redis from './redis';
