+++ bigbluebutton-html5/imports/startup/server/settings.js
